+++ apps/cold-platform-openai/src/app/app.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { NestModule, PrismaModule } from '@coldpbc/nest';
 import { ServeStaticModule } from '@nestjs/serve-static';
+import { BullModule } from '@nestjs/bull';
 
 @Module({})
 export class AppModule {
@@ -8,7 +9,7 @@ export class AppModule {
     return {
       module: AppModule,
       imports: [
-        await NestModule.forRootAsync(),
+        await NestModule.forRootAsync(2),
         ServeStaticModule.forRoot({
           serveStaticOptions: {
             index: false,
@@ -17,6 +18,9 @@ export class AppModule {
           serveRoot: '../../../assets',
         }),
         PrismaModule,
+        BullModule.registerQueue({
+          name: 'openai',
+        }),
       ],
       controllers: [],
       providers: [],