+++ apps/cold-platform-openai/src/main.ts
@@ -1,7 +1,6 @@
 import '../../../libs/nest/src/lib/tracer';
 import { NestFactory } from '@nestjs/core';
-import * as dotenv from 'dotenv';
-// eslint-disable-next-line @nx/enforce-module-boundaries
+import * as dotenv from 'dotenv'; // eslint-disable-next-line @nx/enforce-module-boundaries
 import { AppModule } from './app/app.module';
 import { WorkerLogger } from '@coldpbc/nest';