+++ libs/nest/src/lib/index.ts
@@ -1,3 +1,4 @@
+export * from "./cuid2-generator.service"
 export * from "./nest.module"
 export * from "./tracer"
 export * from "./authorization/index"