+++ libs/nest/src/lib/prisma/prisma.service.ts
@@ -1,11 +1,92 @@
 import { INestApplication, Injectable, OnModuleInit } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
+import { WorkerLogger } from '../worker';
+
+interface LogEvent {
+  timestamp: Date;
+  message: string;
+  target: string;
+}
+
+interface QueryEvent {
+  timestamp: Date;
+  query: string;
+  params: any;
+  duration: number;
+  target: string;
+}
 
 @Injectable()
 export class PrismaService extends PrismaClient implements OnModuleInit {
+  constructor() {
+    super({
+      errorFormat: process.env['NODE_ENV'] === 'development' ? 'minimal' : 'minimal',
+      log: [
+        {
+          emit: 'event',
+          level: 'query',
+        },
+        {
+          emit: 'stdout',
+          level: 'error',
+        },
+        {
+          emit: 'stdout',
+          level: 'info',
+        },
+        {
+          emit: 'stdout',
+          level: 'warn',
+        },
+      ],
+    });
+  }
+
   async onModuleInit() {
     if (process.env['DATABASE_URL']) {
       await this.$connect();
+      const logger = new WorkerLogger('PrismaService');
+
+      // @ts-expect-error Prisma event
+      this.$on('query', (e: QueryEvent) => {
+        if (process.env['NODE_ENV'] !== 'development') {
+          logger.info('Postgres Query', {
+            timestamp: e.timestamp,
+            target: e.target,
+            query: e.query,
+            params: e.params,
+            duration: e.duration,
+          });
+        } else {
+          console.info({
+            message: 'Postgres Query',
+            level: 'info',
+            context: 'PrismaService',
+            timestamp: e.timestamp,
+            target: e.target,
+            query: e.query,
+            params: e.params,
+            duration: e.duration,
+            dd: {
+              service: process.env['DD_SERVICE'],
+              version: process.env['DD_VERSION'],
+              env: process.env['NODE_ENV'],
+            },
+          });
+        }
+      });
+      // @ts-expect-error Prisma event
+      this.$on('info', (e: LogEvent) => {
+        logger.info(e.message, { timestamp: e.timestamp, target: e.target });
+      });
+      // @ts-expect-error Prisma event
+      this.$on('warn', (e: LogEvent) => {
+        logger.warn(e.message, { timestamp: e.timestamp, target: e.target });
+      });
+      // @ts-expect-error Prisma event
+      this.$on('error', (e: LogEvent) => {
+        logger.error(e.message, { timestamp: e.timestamp, target: e.target });
+      });
     }
   }